Centralize duplicated bitcoin simulation logic
Consolidates identical `simulate_bitcoin_prices`, `calculate_moving_averages`,
and `run_trading_algorithm` functions across 10 scripts into a shared
`trading_utils.py` module. Updates imports and removes duplicate code.
